Make non-actors also play spell casting sounds
