+++ apps/openmw/mwmechanics/spellcasting.hpp
@@ -92,6 +92,8 @@ namespace MWMechanics
         /// @note Auto detects if spell, ingredient or potion
         bool cast (const std::string& id);
 
+        void playSpellCastingEffects(const std::string &spellid);
+
         /// @note \a target can be any type of object, not just actors.
         /// @note \a caster can be any type of object, or even an empty object.
         void inflict (const MWWorld::Ptr& target, const MWWorld::Ptr& caster,
